Updated admin passwords. Updated some stuff.

+++ Controllers/AdminController.cs
@@ -5,25 +5,31 @@ using NewsAppServer.Database;
 namespace NewsAppServer.Controllers {
     public class AdminController {
         public AdminController(WebApplication app) {
-            app.MapPost("/admin/login", (HttpContext http, DatabaseManager db,
+            app.MapPost("/admin/login", async (HttpContext http, DatabaseManager db,
                 [FromForm] string adminPassword) => {
 
                 try {
-                    CheckAdmin(db, adminPassword);
+                        bool isAdmin = await CheckAdmin(db, adminPassword);
+                        if (!isAdmin) {
+                            throw new Exception();
+                        }
                     return Results.Ok();
 
                 } catch (Exception) {
-                    return Results.Unauthorized();
+                        return Results.Unauthorized();
                 }
 
             }).DisableAntiforgery()
             .RequireRateLimiting("fixed");
 
-            app.MapPost("/admin/add", (HttpContext http, DatabaseManager db,
+            app.MapPost("/admin/add", async (HttpContext http, DatabaseManager db,
                 [FromForm] string adminPassword, [FromForm] string currentAdmin) => {
 
                     try {
-                        CheckAdmin(db, currentAdmin);
+                        bool isAdmin = await CheckAdmin(db, currentAdmin);
+                        if (!isAdmin) {
+                            throw new Exception();
+                        }
                         db.AddAdminPassword(adminPassword);
                         return Results.Ok();
 
@@ -34,11 +40,14 @@ namespace NewsAppServer.Controllers {
                 }).DisableAntiforgery()
             .RequireRateLimiting("fixed");
 
-            app.MapPost("/admin/remove", (HttpContext http, DatabaseManager db,
+            app.MapPost("/admin/remove", async (HttpContext http, DatabaseManager db,
                 [FromForm] string adminPassword, [FromForm] string currentAdmin) => {
 
                     try {
-                        CheckAdmin(db, currentAdmin);
+                        bool isAdmin = await CheckAdmin(db, currentAdmin);
+                        if (!isAdmin) {
+                            throw new Exception();
+                        }
                         db.RemoveAdmin(adminPassword);
                         return Results.Ok();
 
@@ -49,12 +58,15 @@ namespace NewsAppServer.Controllers {
                 }).DisableAntiforgery()
             .RequireRateLimiting("fixed");
 
-            app.MapPost("/admin/edit", (HttpContext http, DatabaseManager db,
+            app.MapPost("/admin/edit", async (HttpContext http, DatabaseManager db,
                 [FromForm] string oldAdminPassword, [FromForm] string currentAdmin,
                 [FromForm] string newAdminPassword) => {
 
                     try {
-                        CheckAdmin(db, currentAdmin);
+                        bool isAdmin = await CheckAdmin(db, currentAdmin);
+                        if (!isAdmin) {
+                            throw new Exception();
+                        }
                         db.EditAdminPassword(oldAdminPassword, newAdminPassword);
                         return Results.Ok();
 
@@ -66,11 +78,12 @@ namespace NewsAppServer.Controllers {
             .RequireRateLimiting("fixed");
         }
 
-        private async void CheckAdmin(DatabaseManager db, string pass) {
+        private static async Task<bool> CheckAdmin(DatabaseManager db, string pass) {
             List<string> passwords = await db.GetAdminPasswords();
-            if (!passwords.Contains(pass)) {
-                throw new Exception();
-            }
+            Console.WriteLine(passwords.First());
+            Console.WriteLine(pass);
+
+            return passwords.Contains(pass);
         }
     }
 }
